Quick and Dirty: Lazy Man's Opener
Will open the corresponding Plex app/url when movie art is clicked.
This DOES NOT open directly to the movie, as the Plex app (publicly)
doesn’t support this for web or mobile.

This will go to web when browser is over 768px wide (desktop) and will
try to open the plexapp:// when under 768px (mobile). I will handle
edge cases later to make this work correctly for users under 768px on
desktop and over 768px on mobile.